requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.1
Flask-Cors==4.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
neo4j==5.17.0
python-decouple==3.8
pytz==2023.4
Werkzeug==3.0.1
blinker==1.7.0
click==8.1.7
Flask==3.0.1
Flask-Cors==4.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
neo4j==5.17.0
numpy==1.26.3
python-decouple==3.8
pytz==2023.4
Werkzeug==3.0.1
blinker==1.7.0
click==8.1.7
Flask==3.0.1
Flask-Cors==4.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
neo4j==5.17.0
numpy==1.26.3
python-decouple==3.8
python-dotenv==1.0.1
pytz==2023.4
Werkzeug==3.0.1
blinker==1.7.0
click==8.1.7
Flask==3.0.1
Flask-Cors==4.0.0
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
neo4j==5.17.0
numpy==1.26.3
packaging==23.2
python-decouple==3.8
python-dotenv==1.0.1
pytz==2023.4
Werkzeug==3.0.1
